10.0 sale_triple_discount
there are cases where the cache trick done in triple_discount_preprocess
is lost when sale.order.line _compute_amount is called and processes the
1st line of the sale order => all remaining lines are impacted, and don't
have the discounts 2 and 3 taken into account.

Not updating the sale.order _amount_all method because it is unlikely that
fields on the lines depend on the total amount of the sale.