[FIX] sale_coupon_advanced: counter forced reward
Previous fix was not enough after odoo source code was updated, which
still caused to break functionality.

In order for it to work in all cases, instead of removing incorrectly
forced lines, now we check if line can be added and add that line before
it is expected to be used by sale_coupon module.

Because trigger points are different: for applying coupon and
recomputing promotions, had to inject this logic into multiple places.